requirements.txt
pykrx
pandas
numpy
pyarrow
ta
faiss-cpu
fastdtw
tslearn
scikit-learn
fastapi
uvicorn
pydantic
